+++ activation/validation.go
@@ -3,6 +3,7 @@ package activation
 import (
 	"bytes"
 	"fmt"
+
 	"github.com/spacemeshos/go-spacemesh/common/types"
 	"github.com/spacemeshos/go-spacemesh/signing"
 	"github.com/spacemeshos/post/proving"
@@ -28,7 +29,7 @@ func NewValidator(poetDb poetDbAPI, cfg PostConfig) *Validator {
 // Some of the Post metadata fields validation values is ought to eventually be derived from
 // consensus instead of local configuration. If so, their validation should be removed to contextual validation,
 // while still syntactically-validate them here according to locally configured min/max values.
-func (v *Validator) Validate(minerID signing.PublicKey, nipost *types.NIPost, expectedChallenge types.Hash32, numUnits uint) error {
+func (v *Validator) Validate(minerID signing.PublicKey, nipost *types.NIPost, expectedChallenge types.Hash32, numUnits uint64) error {
 	if !bytes.Equal(nipost.Challenge[:], expectedChallenge[:]) {
 		return fmt.Errorf("invalid `Challenge`; expected: %x, given: %x", expectedChallenge, nipost.Challenge)
 	}
@@ -71,7 +72,7 @@ func (v *Validator) Validate(minerID signing.PublicKey, nipost *types.NIPost, ex
 
 // ValidatePost validates a Proof of Space-Time (Post). It returns nil if validation passed or an error indicating why
 // validation failed.
-func (v *Validator) ValidatePost(id []byte, Post *types.Post, PostMetadata *types.PostMetadata, numUnits uint) error {
+func (v *Validator) ValidatePost(id []byte, Post *types.Post, PostMetadata *types.PostMetadata, numUnits uint64) error {
 	p := (*proving.Proof)(Post)
 
 	m := new(proving.ProofMetadata)
